clean up main.go and separation in keysigner package
